increase SQL fieldsize of CostplanItem Number to fit size of Property

+++ HouseBuildingBlog/HouseBuildingBlog.Persistence.MSSql/Scripts/0.3.0/0.3.0.sql
@@ -8,7 +8,7 @@ CREATE TABLE Costplan.CostplanItems (
 	CostplanItemId uniqueidentifier NOT NULL,
 	Name nvarchar(max) NOT NULL,
 	EstimatedCost decimal(10,2) NULL,
-	Number smallint NOT NULL,
+	Number int NOT NULL,
 	CONSTRAINT PK_CostplanItems PRIMARY KEY (CostplanItemId)
 )
 GO
